requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
langchain-google-genai==1.0.2
langgraph==0.0.35
google-genai
groq==0.5.0
tavily-python==0.3.3
beautifulsoup4==4.12.3
requests==2.31.0
aiosqlite==0.20.0
python-dotenv==1.0.1
transformers==4.40.1
torch==2.2.2
Pillow==10.3.0
datasets==2.18.0
duckduckgo-search==5.3.0
pydantic==2.7.0
